backend/jira-git_mcp-server/requirements.txt
python-dotenv
aiohttp
mcp
google-auth
google-genai